requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
dj-email-url==1.0.6
Django==4.2.1
django-cache-url==3.4.4
django-cors-headers==3.14.0
django-cors-middleware==1.5.0
djangorestframework==3.14.0
drf-yasg==1.21.5
environs==8.0.0
gunicorn==19.9.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
permission==0.4.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.30
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==5.1.0
